back/mapsio/requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
Pillow==7.0.0
pip==21.2.4
pyexcel
pyexcel-xlsx
pandas
psycopg2